+++ lib/metronome_web.dart
@@ -44,7 +44,6 @@ class MetronomeWeb extends MetronomePlatform {
     String accentedPath = '',
     int bpm = 120,
     int volume = 50,
-    bool enableSession = true,
     bool enableTickCallback = false,
     int timeSignature = 4,
     int sampleRate = 44100,